[#113] refactor : 나의펫 정보 공통 컴포넌트로 구현
