+++ app/settings/page.css.ts
@@ -1,7 +1,7 @@
 import { style } from "@vanilla-extract/css";
 
 export const container = style({
-  padding: "2rem 1.6rem 1.6rem",
+  padding: "0 1.6rem 1.6rem",
 });
 
 export const listContainer = style({
