+++ app/settings/page.tsx
@@ -21,22 +21,24 @@ const Page = () => {
   };
 
   return (
-    <div className={styles.container}>
+    <>
       <MypetCarousel />
-      <Link href="/profile">
-        <MyProfile />
-      </Link>
-      <div className={styles.listContainer}>
-        <MenuList href="/received-invites" src={HeartIcon} alt="heart icon" text="초대 받은 내역" />
-        <MenuList href="/faq" src={QuestionIcon} alt="question icon" text="FAQ" />
-        <MenuList href="/ask" src={MessageIcon} alt="message icon" text="1:1 문의하기" />
-        <MenuList href="/notice" src={NoticeIcon} alt="notice icon" text="공지사항" />
-        <span className={styles.logout} onClick={openModalFunc}>
-          로그아웃
-        </span>
+      <div className={styles.container}>
+        <Link href="/profile">
+          <MyProfile />
+        </Link>
+        <div className={styles.listContainer}>
+          <MenuList href="/received-invites" src={HeartIcon} alt="heart icon" text="초대 받은 내역" />
+          <MenuList href="/faq" src={QuestionIcon} alt="question icon" text="FAQ" />
+          <MenuList href="/ask" src={MessageIcon} alt="message icon" text="1:1 문의하기" />
+          <MenuList href="/notice" src={NoticeIcon} alt="notice icon" text="공지사항" />
+          <span className={styles.logout} onClick={openModalFunc}>
+            로그아웃
+          </span>
+        </div>
       </div>
       {isModalOpen && <Modal text="로그아웃 하시겠습니까??" buttonText="확인" onClick={handleConfirm} onClose={closeModalFunc} />}
-    </div>
+    </>
   );
 };
