+++ components/Setting/InviteModal/InviteCode/index.tsx
@@ -1,8 +1,8 @@
 import * as styles from "./style.css";
-import card from "./mockData.json";
-import calculateAge from "@/utils/calculateAge";
+import petInfo from "./mockData.json";
 import CopyIcon from "@/assets/copy.svg?url";
 import Image from "next/image";
+import MyPetInfo from "@/components/Setting/MyPetInfo";
 
 const InviteCode = () => {
   const inviteCode = "seul1234";
@@ -23,19 +23,7 @@ const InviteCode = () => {
     <>
       <h3 style={{ fontSize: "1.6rem", fontWeight: "500", marginBottom: "1.1rem" }}>마이펫 초대 코드</h3>
       <section className={styles.codeContainer}>
-        <div
-          className={styles.profile}
-          style={{
-            backgroundImage: `url(${card.petImageUrl})`,
-          }}
-        />
-        <div className={styles.petInfo}>
-          <span style={{ fontSize: "1.4rem", fontWeight: "700" }}>{card.name}</span>
-          <span style={{ fontSize: "1.2rem", fontWeight: "500", color: "var(--Gray81)" }}>{card.breed}</span>
-          <span style={{ fontSize: "1.2rem", fontWeight: "500", color: "var(--Gray81)" }}>
-            {card.gender} · {calculateAge(card.birth)}
-          </span>
-        </div>
+        <MyPetInfo petInfo={petInfo} />
         <div className={styles.copyContainer}>
           <span style={{ fontSize: "1.4rem", fontWeight: "500", color: "var(--Gray81)" }}>{inviteCode}</span>
           <button className={styles.copyIcon} onClick={handleCopyClick}>
